+++ Calculator/Calculator/SwiftCalculatorApp.swift
@@ -8,18 +8,6 @@
 
 import SwiftUI
 
-final class Theme: ObservableObject {
-    @Published var primaryColour = Color(light: Color(.brown), dark: Color(.brown))
-    @Published var accentColour = Color(light: .orange, dark: .orange)
-    @Published var textDisplayFieldForegroundColour = Color(light: .black, dark: .white)
-    @Published var textDisplayFieldBackgroundColour = Color(light: .white, dark: .black)
-    @Published var viewSeparatorColour = Color(light: .black, dark: .white)
-    @Published var buttonForegroundColour = Color(light: .white, dark: .white)
-    
-    static var labelFontToHeightRatio: CGFloat = 0.33
-    static var defaultAnimationDuration = 0.5
-}
-
 @main
 struct SwiftCalculatorApp: App {
     @StateObject var currentTheme = Theme()
